requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile '.\requirements.in'
#
asgiref==3.11.0
    # via django
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
crispy-bootstrap5==2025.6
    # via -r requirements.in
django==6.0.1
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-crispy-forms
    #   djangorestframework
    #   drf-spectacular
    #   drf-spectacular-sidecar
django-crispy-forms==2.5
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-environ==0.12.0
    # via -r requirements.in
djangorestframework==3.16.1
    # via
    #   -r requirements.in
    #   drf-spectacular
drf-spectacular==0.29.0
    # via -r requirements.in
drf-spectacular-sidecar==2026.1.1
    # via -r requirements.in
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.26.0
    # via drf-spectacular
jsonschema-specifications==2025.9.1
    # via jsonschema
pyyaml==6.0.3
    # via drf-spectacular
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
sqlparse==0.5.5
    # via django
tzdata==2025.3
    # via django
uritemplate==4.2.0
    # via drf-spectacular
whitenoise==6.11.0
    # via -r requirements.in
